#117 - reducing StationPlanetary intermodule communications class amount
